+++ src/jerror.rs
@@ -695,7 +695,7 @@ unsafe extern "C" fn output_message(mut cinfo: crate::jpeglib_h::j_common_ptr) {
             .expect("non-null function pointer"),
     )
     .expect("non-null function pointer")(cinfo, buffer.as_mut_ptr());
-      eprintln!("{:}",
+       eprintln!("{:}",
           unsafe {
     std::ffi::CStr::from_ptr(buffer.as_mut_ptr() as
                                  *const libc::c_char).to_str().unwrap()
@@ -722,7 +722,7 @@ unsafe extern "C" fn emit_message(
          * the policy implemented here is to show only the first warning,
          * unless trace_level >= 3.
          */
-        if (*err).num_warnings == 0i32 as libc::c_long || (*err).trace_level >= 3i32 {
+        if (*err).num_warnings == 0i64 || (*err).trace_level >= 3i32 {
             Some((*err).output_message.expect("non-null function pointer"))
                 .expect("non-null function pointer")(cinfo);
         }
@@ -812,7 +812,7 @@ unsafe extern "C" fn format_message(
  */
 
 unsafe extern "C" fn reset_error_mgr(mut cinfo: crate::jpeglib_h::j_common_ptr) {
-    (*(*cinfo).err).num_warnings = 0i32 as libc::c_long;
+    (*(*cinfo).err).num_warnings = 0i64;
     /* trace_level is not reset since it is an application-supplied parameter */
     (*(*cinfo).err).msg_code = 0i32;
     /* may be useful as a flag for "no error" */
@@ -846,7 +846,7 @@ pub unsafe extern "C" fn jpeg_std_error(
     (*err).reset_error_mgr =
         Some(reset_error_mgr as unsafe extern "C" fn(_: crate::jpeglib_h::j_common_ptr) -> ());
     (*err).trace_level = 0i32;
-    (*err).num_warnings = 0i32 as libc::c_long;
+    (*err).num_warnings = 0i64;
     (*err).msg_code = 0i32;
     /* Initialize message table pointers */
     (*err).jpeg_message_table = jpeg_std_message_table.as_ptr(); /* for safety */
